catalog-builder/requirements.txt
requests
xmltodict
pyyaml
retrying
